+++ aesara/tensor/basic_opt.py
@@ -8,3 +8,6 @@ warnings.warn(
 )
 
 from aesara.tensor.rewriting.basic import *  # noqa: F401 E402 F403
+from aesara.tensor.rewriting.elemwise import *  # noqa: F401 E402 F403
+from aesara.tensor.rewriting.extra_ops import *  # noqa: F401 E402 F403
+from aesara.tensor.rewriting.shape import *  # noqa: F401 E402 F403
