+++ aesara/tensor/rewriting/__init__.py
@@ -1,4 +1,7 @@
 import aesara.tensor.rewriting.basic
+import aesara.tensor.rewriting.elemwise
+import aesara.tensor.rewriting.extra_ops
 import aesara.tensor.rewriting.math
+import aesara.tensor.rewriting.shape
 import aesara.tensor.rewriting.subtensor
 import aesara.tensor.rewriting.uncanonicalize
